Added an intermediate class RecyclerViewActivity, for SearchActivity and MainActivity in the hierarchy of activity inheritance
