+++ app/src/main/java/com/ikvych/cocktail/ui/activity/SplashActivity.kt
@@ -3,10 +3,10 @@ package com.ikvych.cocktail.ui.activity
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
-import androidx.appcompat.app.AppCompatActivity
 import com.ikvych.cocktail.R
+import com.ikvych.cocktail.ui.base.BaseActivity
 
-class SplashActivity : AppCompatActivity() {
+class SplashActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
